refactor: move FeedbackNPS component
