+++ src/component/common/util.js
@@ -1,5 +1,4 @@
 import { weightTypes } from '../feature/FeatureView/FeatureVariants/FeatureVariantsList/AddFeatureVariant/enums';
-import differenceInDays from 'date-fns/differenceInDays';
 
 export const filterByFlags = flags => r => {
     if (r.flag && !flags[r.flag]) {
@@ -80,30 +79,3 @@ export const modalStyles = {
         transform: 'translate(-50%, -50%)',
     },
 };
-
-export const showPnpsFeedback = feedbackList => {
-    if (!feedbackList) return;
-    if (feedbackList.length > 0) {
-        const feedback = feedbackList.find(
-            feedback => feedback.feedbackId === PNPS_FEEDBACK_ID
-        );
-
-        if (!feedback) return false;
-
-        if (feedback.neverShow) {
-            return false;
-        }
-
-        if (feedback.given) {
-            const SIX_MONTHS_IN_DAYS = 182;
-            const now = new Date();
-            const difference = differenceInDays(now, new Date(feedback.given));
-
-            return difference > SIX_MONTHS_IN_DAYS;
-        }
-        return false;
-    }
-    return true;
-};
-
-export const PNPS_FEEDBACK_ID = 'pnps';
